Revert "debugs, revert it later"
This reverts commit ac8ea98597ca7c69b7e705279ee569fc2ea70360.

+++ library/source/open_cdm_ext.cpp
@@ -210,7 +210,7 @@ OpenCDMError opencdm_session_get_challenge_data(struct OpenCDMSession *mOpenCDMS
             kLog << error << "Failed to get challenge data - operation returned NOK status";
             return ERROR_FAIL;
         }
-        *challengeSize = challengeVec.size();
+        *challengeSize = static_cast<uint32_t>(challengeVec.size());
         memcpy(challenge, challengeVec.data(), challengeVec.size());
     }
     return ERROR_NONE;
